requirements.txt
# Core Django
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Deprecated==1.2.18
Django==5.2.6
django-filter==25.1
django-mcp-server==0.5.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
mcp==1.13.1
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-sqlite3==0.43b0
opentelemetry-instrumentation-tortoiseorm==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
protobuf==4.25.8
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.3
zipp==3.23.0
drf-spectacular==0.27.2

# Environment and Configuration
python-dotenv==1.1.1

# Optional: Redis cache backend (uncomment if needed)
# django-redis==5.4.0

# Optional: JSON logging for production (uncomment if needed)
# python-json-logger==2.0.7

# Optional: Development tools (uncomment for development)
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3

# Optional: Code quality tools (uncomment for development)
# black==23.12.1
# pylint==3.0.3
# flake8==7.0.0
# bandit==1.7.6

# OpenTelemetry instrumentation
opentelemetry-instrumentation-openai
